stroke patient management/requirements.txt
Flask==2.2.5
mysql-connector-python==8.1.0
pymongo==4.4.0
werkzeug==2.2.3
dnspython==2.4.2